refactor : feature comment 폴더명 변경
