+++ src/pages/PostsManagerPage.tsx
@@ -2,9 +2,9 @@ import { PostsHeader } from "@widgets/post/ui"
 import { PostsTableContainer } from "@widgets/post/ui/PostsTableContainer"
 import { UserDetailModal } from "@features/user-management/ui/UserDetailModal"
 import { PostDetailModal } from "@features/post-management/ui/PostDetailModal"
-import EditCommentModal from "@features/comment-management/ui/EditCommentModal"
+import EditCommentModal from "@features/comment/ui/EditCommentModal"
 import EditPostModal from "@features/post-management/ui/EditPostModal"
-import CreateCommentModal from "@features/comment-management/ui/CreateCommentModal"
+import CreateCommentModal from "@features/comment/ui/CreateCommentModal"
 import CreatePostModal from "@features/post-management/ui/CreatePostModal"
 import { useURLSync } from "@features/filter-management/model/useURLSync"
